refactor: move vue.ts to vue/store.ts
