+++ src/adapters/devtool-plugin.ts
@@ -1,7 +1,7 @@
 import { BA0, BG, BG0, BM0 } from '../core/base'
 
 import { Dictionary } from '../utils'
-import { VueStore } from '../vue'
+import { VueStore } from '../vue/store'
 import { flattenGetters } from './vuex'
 
 const devtoolHook =
