+++ src/index.ts
@@ -1,4 +1,4 @@
 export { module } from './core/module'
 export { inject, Getters, Mutations, Actions } from './core/base'
 export { convertVuexPlugin } from './adapters/vuex'
-export { store, install } from './vue'
+export { store, install } from './vue/store'
